move AverageAggregationResult implementation into own file
